requirements.txt
altgraph==0.17.4
asttokens==2.4.1
backcall==0.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
dlib @ file:///D:/workSpace/face_detection_project/dlib-19.22.99-cp38-cp38-win_amd64.whl#sha256=b02321413b8765159ae552967ab9781a981324b2a389924bdb8c4ccab227f160
et-xmlfile==1.1.0
executing==2.0.1
facenet-pytorch==2.6.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
ipykernel==6.29.5
ipython==8.12.3
jalali_core==1.0.0
jdatetime==5.0.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.1
numpy==1.24.4
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.0.3
parso==0.8.4
pefile==2023.2.7
pickle-mixin==1.0.2
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure-eval==0.2.2
Pygments==2.18.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==26.0.3
requests==2.32.3
scipy==1.10.1
six==1.16.0
stack-data==0.6.3
sympy==1.13.0
torch==2.2.2
torchvision==0.17.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
ttkbootstrap==1.10.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
zipp==3.19.2
